+++ application/modules/editor/database/sql-changelog-7.35.2-2026-03-03.sql
@@ -1,0 +1,35 @@
+
+-- /*
+-- START LICENSE AND COPYRIGHT
+--
+--  This file is part of translate5
+--
+--  Copyright (c) 2013 - 2017 Marc Mittag; MittagQI - Quality Informatics;  All rights reserved.
+--
+--  Contact:  http://www.MittagQI.com/  /  service (ATT) MittagQI.com
+--
+--  This file may be used under the terms of the GNU AFFERO GENERAL PUBLIC LICENSE version 3
+--  as published by the Free Software Foundation and appearing in the file agpl3-license.txt
+--  included in the packaging of this file.  Please review the following information
+--  to ensure the GNU AFFERO GENERAL PUBLIC LICENSE version 3 requirements will be met:
+--  http://www.gnu.org/licenses/agpl.html
+--
+--  There is a plugin exception available for use with this release of translate5 for
+--  translate5: Please see http://www.translate5.net/plugin-exception.txt or
+--  plugin-exception.txt in the root folder of translate5.
+--
+--  @copyright  Marc Mittag, MittagQI - Quality Informatics
+--  @author     MittagQI - Quality Informatics
+--  @license    GNU AFFERO GENERAL PUBLIC LICENSE version 3 with plugin-execption
+--              http://www.gnu.org/licenses/agpl.html http://www.translate5.net/plugin-exception.txt
+--
+-- END LICENSE AND COPYRIGHT
+-- */
+
+-- userGroup calculation: basic: 1; editor: 2; pm: 4; admin: 8
+            
+INSERT INTO `LEK_change_log` (`dateOfChange`, `jiraNumber`, `type`, `title`, `description`, `userGroup`) VALUES ('2026-03-03', 'TRANSLATE-5034', 'change', 't5memory - Filter TMX on import', '7.35.2: conversion script did fail for empty TMs
+7.35.0: Apply filters on TMX on import time from translate5 side
+All Translate5 TMs in instance will be marked as "not converted" and matches may have lower matches.
+Language resources affected by this should be converted manually.', '15'),
+('2026-03-03', 'TRANSLATE-5292', 'bugfix', 'MatchAnalysis & Pretranslation - 104 match faulty', 'FIXED: sublanguage penalties detection problem for TermCollections, caused by usage of terms having not the best matching languages possible', '15');